libhbalinux: Return model string for HBA_GetAdapterAttributes
The model string of HBA_GetAdapterAttributes is currently
always returned as "Unknown". This patch is to return a better
meaningful string as the model string.

If the device is a newly developed product, and the model description
has not been added into pci.ids yet, use the model description
constructed for new product as the model string.

If the device name (model description) has already been added into
the pci.ids file, use the first word of the model description as
the model string. If the space after the first word is not found
(new product), use the entire model description as the model string.

Signed-off-by: Steve Ma <steve.ma@intel.com>