libhbalinux: Fix build warnings
Fix a build warning of strnlen:
  lport.c: In function ‘sysfs_scan’:
  lport.c:390: warning: implicit declaration of function ‘strnlen’

The build warning is because __USE_GNU is not defined in
/usr/include/string.h.  The man page of strnlen requires
#define _GNU_SOURCE to present before #include <string.h>.
The fix is to make #define _GNU_SOURCE to be the very first
thing in the file utils.h which is the first include file of
lport.c, and because almost every header file in /usr/include
includes features.h, __USE_GNU will be defined by features.h
if _GNU_SOURCE is defined. We do not use #define __USE_GNU
before #include <string.h> because __USE_GNU is a glibc
internal macro.

In addition to this fix, an unused variable is also removed.

Signed-off-by: Steve Ma <steve.ma@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>