libhbalinux: modify to match latest kernel /sys
libhbalinux expected that /sys/class/fc_hosts/<hostx> would be a symlink.
That will be a directory, and readlink(2) fails for those.

Change to look for the symlink "/sys/class/fc_hosts/<hostx>/device".

Don't expect the string "/virtual/net/" to be in the symlink.
If it was at one time, it no longer is in current kernels.

To find the real interface for a vlan, compare the iflink and ifindex.
If they are not the same, look for the interface who's ifindex matches
the iflink of the vlan.

Signed-off-by: Joe Eykholt <jeykholt@cisco.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>