libhbalinux: NumberOfPorts per Adapter count correction
The NumberOfPorts attribute for an adapter was set based on whether
the PCI device is a single function or multi-function device. If the
device was multi-function then the value of NumberOfPorts attribute
was set to 2 else it was set to 1.

This does not seem to have any premise based on actual number of ports
so for now each adapter by default will have only 1 port.

Signed-off-by: Neerav Parikh <Neerav.Parikh@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>