libhbalinux: Support for SCSI READ CAPACITY (16)
This patch adds support for sending SCSI READ
CAPACITY (16) via libhbalinux.

The decision to send SCSI READ CAPACITY (16) or
SCSI READ CAPACITY (10) is based on the size of
the buffer being sent to the API by the calling
application. If the buffer size is big enough
to hold SCSI READ CAPACITY (16) response data
the sg_issue_read_capacity() will issue that
command else SCSI READ CAPACITY (10) will be
sent to the device.

Signed-off-by: Neerav Parikh <Neerav.Parikh@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: Yi Zou <yi.zou@intel.com>